Добавляет тесты на редьюсеры
