+++ src/store/reducer/locations/locations.js
@@ -15,4 +15,4 @@ const reducer = (state = initialState, action) => {
   return state;
 };
 
-export {reducer};
+export {reducer, initialState};
