+++ src/store/reducer/places/places.js
@@ -17,4 +17,4 @@ const reducer = (state = initialState, action) => {
   return state;
 };
 
-export {reducer};
+export {reducer, initialState};
