+++ src/store/reducer/reviews/reviews.js
@@ -14,4 +14,4 @@ const reducer = (state = initialState, action) => {
   return state;
 };
 
-export {reducer};
+export {reducer, initialState};
